Remove duplicate hotkeys code
